gender

+++ HRIS/Hris.Web.Main/ClientApp/app/components/shared/http.client.ts
@@ -1,36 +1,53 @@
 ﻿import {Injectable} from "@angular/core";
-import { Http, Headers } from "@angular/http";
+import { Http, Response } from "@angular/http";
 import { ShareModel } from "./share.model";
 
 import "rxjs/add/operator/finally";
+import { Observable } from "rxjs/Observable";
 
 @Injectable()
 export class HttpClient {
+
+    private readonly observable: Observable<Response>;
     constructor(private readonly http: Http,
-        private readonly sm: ShareModel) { }
+        private readonly sm: ShareModel) {
 
-    get(url: string) {
-        return this.http.get(url,
-            {
-                headers: this.sm.createAuthorizationHeader()
-            }).finally(() => { this.sm.isAjaxProcessing = false; });
+        this.observable = new Observable<Response>(() => {
+
+        });
     }
 
-    post(url: string, body: any) {
-        return this.http.post(url,
-            body,
-            {
-                headers: this.sm.createAuthorizationHeader()
-            }).finally(() => { this.sm.isAjaxProcessing = false; });
+    get(url: string | undefined) {
+        if (url)
+            return this.http.get(url,
+                {
+                    headers: this.sm.createAuthorizationHeader()
+                }).finally(() => { this.sm.isAjaxProcessing = false; });
+
+        else return this.observable;
     }
 
-    put(url: string, body: any) {
-        return this.http.put(url, body, { headers: this.sm.createAuthorizationHeader() })
-            .finally(() => { this.sm.isAjaxProcessing = false; });
+    post(url: string | undefined, body: any) {
+        if (url)
+            return this.http.post(url,
+                body,
+                {
+                    headers: this.sm.createAuthorizationHeader()
+                }).finally(() => { this.sm.isAjaxProcessing = false; });
+        else return this.observable;
     }
 
-    delete(url: string) {
-        return this.http.delete(url, { headers: this.sm.createAuthorizationHeader() })
-            .finally(() => { this.sm.isAjaxProcessing = false });
+    put(url: string | undefined, body: any) {
+        if (url)
+            return this.http.put(url, body, { headers: this.sm.createAuthorizationHeader() })
+                .finally(() => { this.sm.isAjaxProcessing = false; });
+        else return this.observable;
+    }
+
+    delete(url: string | undefined) {
+        if (url)
+            return this.http.delete(url, { headers: this.sm.createAuthorizationHeader() })
+                .finally(() => { this.sm.isAjaxProcessing = false });
+        else return this.observable;
     }
 }
